Move util and roleUtils into separate utils folder
