+++ commands/user/clear.ts
@@ -4,7 +4,7 @@ import { ICommand } from "wokcommands";
 import { classModel } from "../../models/classModel";
 import { staffModel } from "../../models/staffModel";
 import { yearModel } from "../../models/yearModel";
-import { removeRole } from "../../rolesOps";
+import { removeRole } from "../../utils/roleUtils";
 
 export default {
   name: "clear",
